Separate view concerns from controller over Snips
For performance and against the idiomatic React way, snips want to just
control whether or not they're highlighted and other local changes. This
is because the performance of trying to set these from above is very poor
given the size of the code tree. Instead, we pass the props up from snips
to the parent (CodeBlock) which manages scrolling and tooltips without
needing to re-render the tree. This required some futzing with purity
and encapsulation until I could minimize the dependency set of a wrapper
component (now named CodeBlock -- what was once CodeBlock is now
MainContent) to change only when the source or snips changed
significantly.
