+++ web/js-src/dummies.d.ts
@@ -1,0 +1,2 @@
+declare module 'highlight.js/lib/core';
+declare module 'highlight.js/lib/languages/haskell'
