+++ web/package.json
@@ -4,6 +4,7 @@
   },
   "dependencies": {
     "@flatten-js/interval-tree": "^1.0.11",
+    "@types/highlight.js": "^9.12.4",
     "highlight.js": "^10.1.2",
     "immutable": "^4.0.0-rc.12",
     "mnemonist": "^0.38.0",
