+++ web/public/css/index.css
@@ -129,18 +129,22 @@ ul.flatlist, ol.flatlist {
 							
 						}
 				
-				.snip-1 {
+				span[class*='snip-2-'], span[class*='snip-3-'], span[class*='snip-4-'], span[class*='snip-5-'], span[class*='snip-6-'], span[class*='snip-7-'], span[class*='snip-8-'] {
+					box-shadow:rgba(0, 255, 0, 1.0) 0 1px 0;
+				}
+				span[class*='snip-0-']:hover, span[class*='snip-1-']:hover, span[class*='snip-2-']:hover, .snip.hoverable:hover .snip.focused {
+					background-color:red;
+				}
+				
+				.snip-0-1, .snip-1-1, .snip-2-1 {
 					background-color:rgba(0, 255, 0, 0.2);
 				}
-				.snip-2 {
+				.snip-0-2, .snip-1-2, .snip-2-2 {
 					background-color:rgba(0, 255, 0, 0.4);
 				}
-				.snip-3 {
+				.snip-0-3, .snip-1-3, .snip-2-3 {
 					background-color:rgba(0, 255, 0, 0.6);
 				}
-				.snip-4 {
+				.snip-0-4, .snip-1-4, .snip-2-4 {
 					background-color:rgba(0, 255, 0, 0.8);
-				}
-				.snip.focused {
-					background-color:red;
 				}
